pyproject.toml
[project]
name = "agent_pagerduty"
version = "0.1.0"
license = "Apache-2.0"
description = "A PagerDuty natural language agent using LangChain, LangGraph, and MCP."
readme = "README.md"
authors = [
    { name = "Rehan", email = "rehagraw@cisco.com" },
]
maintainers = [
    { name = "Rehan", email = "rehagraw@cisco.com" },
]
requires-python = ">=3.13,<4.0"
dependencies = [
    "a2a-python>=0.0.1",
    "a2a-sdk>=0.2.1",
    "agentevals>=0.0.7",
    "click>=8.2.0",
    "fastmcp>=2.0.0",
    "langchain-anthropic>=0.3.13",
    "langchain-core>=0.3.60",
    "langchain-google-genai>=2.1.4",
    "langchain-mcp-adapters>=0.1.0",
    "langchain-openai>=0.3.17",
    "langgraph>=0.4.5",
    "pytest>=8.3.5",
    "tabulate>=0.9.0",
    "uv",
    "rich>=14.0.0",
    "sseclient>=0.0.27",
    "fastapi>=0.115.11",
    "uvicorn>=0.34.2",
    "httpx>=0.27.0",
    "httpx-sse>=0.3.0",
    "python-dotenv>=1.0.1",
    "cnoe-agent-utils>=0.1.4",
    "langchain>=0.1.0",
    "pydantic>=2.0.0",
    "a2a>=0.0.1",
]

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.poetry.scripts]
agent_pagerduty_a2a = "agent_pagerduty.protocol_bindings.a2a_server:main"

[tool.poetry.group.dev.dependencies]
pytest-asyncio = "^1.0.0"
pytest-cov = "^6.1.1"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
# Allow lines to be as long as 140.
line-length = 140
# Number of spaces per indentation level
indent-width = 2

[tool.ruff.lint]
select = [
    # pycodestyle
    "E",
    # Pyflakes
    "F",
]
ignore = ["F403"]

[tool.pytest.ini_options]
markers = [
    "unit: marks tests as unit tests",
    "integration: marks tests as integration tests",
    "parsers: marks tests as parser tests"
] 